chore: move sdk logic to logic dir
